Remove a lot of methods from *TypeMethods
